メッセージフォーカスを相対的に動かすショートカットキーの追加 refs 67
git-svn-id: svn://toshia.dip.jp/mikutter/trunk@229 03aab468-d3d2-4883-8b12-f661bbf03fa8

+++ core/addon/openimg.rb
@@ -36,7 +36,7 @@ Module.new do
       if not(url) or not(url.respond_to?(:to_s))
         Delayer.new{
           if cancel
-            w.destroy
+            w.destroy unless w.destroyed?
             cancel.call
           else
             w.set_title("URLの取得に失敗") end }
